Move the feature button to the notes area
